+++ packages/parser/demo-parser.js
@@ -23,7 +23,7 @@ const template = `
         <ul>
           <% for (let tag of post.tags): %>
             <li>
-              <a href="/tag/<%= tag %>" $onclick="handle_tag_click">
+              <a href="/tag/<%= tag %>" @click=this.handle_tag_click>
                 #<%= tag %>
               </a>
             </li>
@@ -43,7 +43,7 @@ const template = `
           </div>
         <% endfor; %>
         
-        <form $onsubmit="add_comment">
+        <form @submit=this.add_comment>
           <textarea $id="comment_text" placeholder="Add a comment..."></textarea>
           <button type="submit">Post Comment</button>
         </form>